backend/backend/requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
python-decouple>=3.8

# Production server & deployment
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
psycopg2-binary>=2.9
